feat: Add OpenShift 4 Disconnected Helper workflow
Add a new workflow called "OpenShift 4 Disconnected Helper" to the repository. This workflow allows users to perform actions on a target server using self-hosted runners. It includes options to download to TAR, push TAR to a registry, and set up Quay and Harbor registries. The workflow consists of two jobs: "configure-kcli-profiles" and "ocp4-disconnected-helper". The jobs are executed on self-hosted runners and have dependencies on each other. The "configure-kcli-profiles" job configures kcli profiles, while the "ocp4-disconnected-helper" job deploys a virtual machine on KVM using the kcli-pipelines repository.