feat: Add GUID input parameter to configure-ocp4-disconnected-helper workflow
Add a new input parameter guid to the configure-ocp4-disconnected-helper workflow. This parameter is required and allows users to specify a GUID. The default value is an empty string.

Refactor deploy.sh script in ocp4-disconnected-helper

Remove unnecessary lines of code in the deploy.sh script in the ocp4-disconnected-helper directory. The lines related to setting environment variables and deploying a VM have been removed.

Update domain variable in deploy.sh script

Update the domain variable in the deploy.sh script to include the GUID. The domain is used for downloading images from the harbor registry.